Adding MotionController into the current ARGUS architecture
